move proxy call shell test into proxy tests
